ответ на запоздалое замечание по реестру
